Squashed commit of the following:
commit 24f11f40359fa05b710c77934880596f85c7f133
Author: sumit <sumit@factors.ai>
Date:   Wed Jan 3 11:40:10 2024 +0530

    resolve conflicts

commit 4f2f9ec46119d8004b33875d336199acf6563d4e
Merge: 0e173fb895 a01b766b5d
Author: sumit <sumit@factors.ai>
Date:   Wed Jan 3 11:39:35 2024 +0530

    Merge branch 'master' into billing-stag-fixes-11d

commit 0e173fb895ef42608b881964f556c9d25b7f711b
Author: sumit <sumit@factors.ai>
Date:   Wed Jan 3 11:18:19 2024 +0530

    dont create customer for invited agents

commit bd57bae876ce9671f6889925dfea4d416acd11aa
Author: sumit <sumit@factors.ai>
Date:   Tue Dec 26 15:48:54 2023 +0530

    change amount type to float to support cents

commit 03b3984ac89f39e0d2104385cbcda1591612ebe1
Merge: 5d7596ed89 169f509419
Author: sumit <sumit@factors.ai>
Date:   Thu Dec 21 13:36:34 2023 +0530

    Merge branch 'master' into billing-stag-fixes-11d

commit 5d7596ed89c54dd01b810e1234dbe21249e30901
Author: sumit <sumit@factors.ai>
Date:   Tue Dec 12 13:19:19 2023 +0530

    format prices and url fixes

+++ backend/src/factors/model/model/billing.go
@@ -28,19 +28,19 @@ func GetNumberOfAccountsForAddOnID(addOnID string) int {
 }
 
 type SubscriptionProductPrice struct {
-	Type         string `json:"type"`
-	Name         string `json:"name"`
-	ExternalName string `json:"external_name"`
-	ID           string `json:"id"`
-	Price        int64  `json:"price"`
-	PeriodUnit   string `json:"period_unit"`
+	Type         string  `json:"type"`
+	Name         string  `json:"name"`
+	ExternalName string  `json:"external_name"`
+	ID           string  `json:"id"`
+	Price        float64 `json:"price"`
+	PeriodUnit   string  `json:"period_unit"`
 }
 
 type DifferentialPrice struct {
-	ID           string `json:"id"`
-	ItemPriceID  string `json:"item_price_id"`
-	ParentItemID string `json:"parent_item_id"`
-	Price        int64  `json:"price"`
+	ID           string  `json:"id"`
+	ItemPriceID  string  `json:"item_price_id"`
+	ParentItemID string  `json:"parent_item_id"`
+	Price        float64 `json:"price"`
 }
 
 type Subscription struct {
@@ -51,10 +51,10 @@ type Subscription struct {
 }
 
 type SubscriptionDetail struct {
-	Type         string `json:"type"`
-	ID           string `json:"id"`
-	Amount       int64  `json:"amount"`
-	ExternalName string `json:"external_name"`
+	Type         string  `json:"type"`
+	ID           string  `json:"id"`
+	Amount       float64 `json:"amount"`
+	ExternalName string  `json:"external_name"`
 }
 
 type UpdateSubscriptionParams struct {
@@ -70,9 +70,9 @@ type AddOnsUpdate struct {
 type Invoice struct {
 	ID          string    `json:"id"`
 	BillingDate time.Time `json:"billing_date"`
-	Amount      int64     `json:"amount`
-	AmountPaid  int64     `json:"amount_paid"`
-	AmountDue   int64     `json:"amount_due`
+	Amount      float64   `json:"amount`
+	AmountPaid  float64   `json:"amount_paid"`
+	AmountDue   float64   `json:"amount_due`
 	Items       []string  `json:"items"`
 }
